Change connectionId from Long to Int

+++ eip-core/src/main/scala/com/digitalpetri/ethernetip/cip/CipConnection.scala
@@ -3,8 +3,8 @@ package com.digitalpetri.ethernetip.cip
 import scala.concurrent.duration.Duration
 
 
-case class CipConnection(o2tConnectionId: Long,
-                         t2oConnectionId: Long,
+case class CipConnection(o2tConnectionId: Int,
+                         t2oConnectionId: Int,
                          serialNumber: Int,
                          originatorVendorId: Int,
                          originatorSerialNumber: Long,
